requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
pymupdf==1.23.8
python-docx==0.8.11
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
tiktoken==0.5.1
email-reply-parser==0.5.12
markdown==3.5.1
beautifulsoup4==4.12.2
aiofiles==23.2.0
httpx==0.25.2
Pillow==10.1.0
